requirements.txt
alembic==1.5.2
ali-open-api==1.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
avalon==0.2.8
backcall==0.2.0
Beaker==1.11.0
bleach==3.2.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
collie==21.6.1
crcmod==1.7
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
demjson==2.2.4
EigenConfigLibrary==0.0.1
elasticsearch==7.10.1
entrypoints==0.3
et-xmlfile==1.1.0
fast-requests==0.0.14
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gevent==21.1.2
greenlet==1.0.0
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
jsonlines==1.2.0
jsonpath==0.82
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kafka-python==2.0.2
kiwisolver==1.3.1
lru-dict==1.1.7
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.12.1
matplotlib==3.4.2
mistune==0.8.4
more-itertools==8.8.0
munkres==1.1.2
mysql-connector-python==8.0.5
mysql-replication==0.25
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.21.0
openpyxl==3.0.1
oss2==2.3.4
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parsel==1.5.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.14
ptyprocess==0.7.0
py==1.10.0
pyahocorasick==1.4.0
pycparser==2.20
pydantic==1.8.2
PyDbLite==3.0.4
PyExecJS==1.5.1
Pygments==2.7.4
pymongo==3.11.2
PyMySQL==0.9.3
pyparsing==2.4.7
pypinyin==0.35.4
pyrsistent==0.17.3
pytest==5.1.2
python-consul==1.1.0
python-dateutil==2.8.1
python-editor==1.0.4
python-gitlab==2.8.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
quota-sdk==0.0.2
readability-lxml==0.7.1
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
retrying==1.3.3
saber-sdk==0.0.34
seer-sdk==0.0.2
Send2Trash==1.5.0
sheep==0.1.14
simplex-sdk==0.0.20
six==1.15.0
SQLAlchemy==1.3.22
tenacity==5.0.3
terminado==0.9.2
testpath==0.4.4
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
XlsxWriter==1.4.3
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
